made a page that shows a table of all the users waitiing to be approved to be EA
