+++ EmpathyKick.csproj
@@ -40,7 +40,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Data\" />
     <Folder Include="Views\Admin\" />
   </ItemGroup>
