+++ Views/Home/PendingEAView.cshtml
@@ -1,0 +1,51 @@
+﻿@using EmpathyKick.Data
+@using EmpathyKick.Models
+@{
+    ViewData["Title"] = "Pending Empathy Admins";
+    var UsersRequestingEAship = Model;
+    User[] h = new User[8];
+    int g = h.Length;
+
+}
+<h2>@ViewData["Title"]</h2>
+@if (UsersRequestingEAship != null && UsersRequestingEAship.Length > 0)
+{
+    <table class="table">
+        <thead>
+            <tr>
+                <th>UserId</th>
+                <th>Username</th>
+                <th>First Name</th>
+                <th>Last Name</th>
+                <th>Action</th>
+            </tr>
+        </thead>
+        <tbody>
+            @foreach (var user in UsersRequestingEAship)
+            {
+                <tr>
+                    <td>@user.UserId</td>
+                    <td>@user.Username</td>
+                    <td>@user.FirstName</td>
+                    <td>@user.LastName</td>
+                    <td>
+                        <button id="approveButton_@user.UserId" class="btn btn-primary" onclick="approveUser(@user.UserId)">Approve</button>
+                    </td>
+                </tr>
+            }
+        </tbody>
+    </table>
+}
+else
+{
+    <p>No pending empathy admins.</p>
+}
+
+@section Scripts {
+    <script>
+        function approveUser(userId) {
+
+            alert("Approved user with UserId: " + userId);
+        }
+    </script>
+}
