+++ Views/Shared/_Layout.cshtml
@@ -36,6 +36,9 @@
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="EATableSelectionView">View Data</a>
                         </li>
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="PendingEAView">Pending EAs</a>
+                        </li>
                         }
                     </ul>
                     @if (@Context.Session.GetString("Username") == null || @Context.Session.GetString("Username") == "")
