+++ appsettings.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": "Server=tcp:empathykick.database.windows.net,1433;Initial Catalog=EmpathyKickDB;Persist Security Info=False;User ID=empathyadmin;Password=Adminpassword123;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
+    "DefaultConnection": "Server=tcp:empathykick.database.windows.net,1433;Initial Catalog=EmpathyKickDB;Persist Security Info=False;User ID=empathyadmin;Password=Adminpassword123;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=120;"
   }
 }
